HW_lesson_2.6 initial commit
