+++ src/main/java/pro/sky/employeewebservice/controller/EmployeeController.java
@@ -1,9 +1,13 @@
-package pro.sky.employeewebservice;
+package pro.sky.employeewebservice.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import pro.sky.employeewebservice.data.Employee;
+import pro.sky.employeewebservice.service.EmployeeService;
+
+import java.util.List;
 
 @RestController
 @RequestMapping("/employee")
@@ -27,5 +31,10 @@ public class EmployeeController {
     @GetMapping("/find")
     public Employee findEmployeeInList(@RequestParam String firstName, @RequestParam String lastName) {
         return employeeService.findEmployee(firstName, lastName);
+    }
+
+    @GetMapping("/employeelist")
+    public List<Employee> printEmployeeList() {
+        return employeeService.getEmployeeList();
     }
 }
