+++ src/main/java/pro/sky/employeewebservice/data/Employee.java
@@ -1,4 +1,4 @@
-package pro.sky.employeewebservice;
+package pro.sky.employeewebservice.data;
 
 import java.util.Objects;
