+++ src/main/java/pro/sky/employeewebservice/exceptions/ArrayIsFullException.java
@@ -1,4 +1,4 @@
-package pro.sky.employeewebservice;
+package pro.sky.employeewebservice.exceptions;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
