+++ src/main/java/pro/sky/employeewebservice/service/EmployeeService.java
@@ -1,7 +1,12 @@
-package pro.sky.employeewebservice;
+package pro.sky.employeewebservice.service;
+
+import pro.sky.employeewebservice.data.Employee;
+
+import java.util.List;
 
 public interface EmployeeService {
     Employee addEmployee(String firstName, String lastName);
     Employee removeEmployee(String firstName, String lastName);
     Employee findEmployee(String firstName, String lastName);
+    List<Employee> getEmployeeList();
 }
