+++ src/main/java/pro/sky/employeewebservice/service/impl/EmployeeServiceImpl.java
@@ -1,0 +1,51 @@
+package pro.sky.employeewebservice.service.impl;
+
+import org.springframework.stereotype.Service;
+import pro.sky.employeewebservice.data.Employee;
+import pro.sky.employeewebservice.exceptions.ArrayIsFullException;
+import pro.sky.employeewebservice.exceptions.EmployeeExistsException;
+import pro.sky.employeewebservice.exceptions.EmployeeNotFoundException;
+import pro.sky.employeewebservice.service.EmployeeService;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Service
+public class EmployeeServiceImpl implements EmployeeService {
+    List<Employee> employees = new ArrayList<>();
+
+    @Override
+    public Employee addEmployee(String firstName, String lastName) {
+        Employee employee = new Employee(firstName, lastName);
+        if (employees.contains(employee)) {
+            throw new EmployeeExistsException();
+        }
+        employees.add(employee);
+        return employee;
+    }
+
+    @Override
+    public Employee removeEmployee(String firstName, String lastName) {
+        Employee employee = new Employee(firstName, lastName);
+        if (!employees.contains(employee)) {
+            throw new EmployeeNotFoundException();
+        }
+        employees.remove(employee);
+        return employee;
+    }
+
+    @Override
+    public Employee findEmployee(String firstName, String lastName) {
+        Employee employee = new Employee(firstName, lastName);
+        if (!employees.contains(employee)) {
+            throw new EmployeeNotFoundException();
+        }
+        return employee;
+    }
+
+    @Override
+    public List<Employee> getEmployeeList() {
+        return employees;
+    }
+
+}
